+++ src/main/java/example/themes/console/ByteAreaColorProvider.java
@@ -54,7 +54,7 @@ public final class ByteAreaColorProvider implements IByteColorProvider
 
     @NotNull
     @Override
-    public Color getRowElementForeground(final int offset, final int rowIndex, final int elementIndex)
+    public Color getRowElementForeground(final long offset, final int rowIndex, final int elementIndex)
     {
         return rowElementForeground;
     }
